movie-review-backend/requirements.txt
Flask>=2.0
Flask-Cors>=3.0
Flask-SQLAlchemy>=3.0
PyMySQL>=1.0
python-dotenv>=1.0
mysql-connector-python>=8.0
werkzeug>=2.0
PyJWT>=2.8
requests>=2.25
